requirements.txt
Django==3.1.3
gunicorn==20.0.4
djangorestframework==3.12.2
Markdown==3.3.3
django-filter==2.4.0
